Enhancement: separate into modules
