+++ package.json
@@ -7,8 +7,7 @@
     "build": "next build",
     "dev": "next dev",
     "lint": "next lint",
-    "start": "next start",
-    "startWebCrawler:dev": "node --loader ts-node/esm --env-file=.env src/server/worker/WebCrawler.ts"
+    "start": "next start"
   },
   "dependencies": {
     "@auth/drizzle-adapter": "^0.7.0",
@@ -38,7 +37,6 @@
   "devDependencies": {
     "@types/eslint": "^8.56.2",
     "@types/node": "^20.11.20",
-    "@types/pg": "^8.11.5",
     "@types/react": "^18.2.57",
     "@types/react-dom": "^18.2.19",
     "@types/uuid": "^9.0.8",
@@ -53,7 +51,6 @@
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "tailwindcss": "^3.4.1",
-    "ts-node": "^10.9.2",
     "typescript": "^5.4.2"
   },
   "ct3aMetadata": {
